+++ src/model/mod.rs
@@ -2,8 +2,8 @@
 pub struct Item {
     pub nome: String,
     pub quantidade: u64,
-    pub preco_unitario: u64,
-    pub valor_total: u64,
+    pub preco_unitario: f64,
+    pub valor_total: f64,
     pub is_resource: bool,
     pub preco_input: String,
 }
